add new parm: worldPositionStays in  Transform::SetParent. update Rigidbody::Start and Collider::Start GameObject::GetComponent<T> now may return subclass.

+++ Engine/Source/Editor/UI/UISlider.hpp
@@ -19,7 +19,7 @@ public:
     bool CheckUpdate(std::string const & label, float & value);
 
 private Q_SLOTS:
-    void OnSpinBoxValueChanged(double v);
+    void OnSpinBoxValueChanged(float v);
 
     void OnSliderValueChanged(int v);
